+++ src/Fragment/Selection.php
@@ -16,6 +16,9 @@ namespace League\Csv\Fragment;
 use League\Csv\FragmentNotFound;
 use const FILTER_VALIDATE_INT;
 
+/**
+ * @internal Internal representation of an Expression Selection.
+ */
 final class Selection
 {
     private const REGEXP_ROWS_COLUMNS_SELECTION = '/^(?<start>\d+)(-(?<end>\d+|\*))?$/';
@@ -29,25 +32,37 @@ final class Selection
         )?
     $/x';
 
-    public static function fromUnknown(): self
-    {
-        return new self(-1, null, -1, null);
-    }
-
     private function __construct(
         public readonly int $rowStart,
         public readonly ?int $rowEnd,
         public readonly int $columnStart,
         public readonly ?int $columnEnd,
-    ) {
-    }
+    ) {}
 
-    public function rowCount(): ?int
+    public function toString(): string
     {
+        if (-1 === $this->columnStart) {
+            return match ($this->rowEnd) {
+                null => ($this->rowStart + 1).'-*',
+                $this->rowStart => (string) ($this->rowStart + 1),
+                default => ($this->rowStart + 1).'-'.($this->rowEnd + 1),
+            };
+        }
+
+        if (-1 === $this->rowStart) {
+            return match ($this->columnEnd) {
+                -1 => ($this->columnStart + 1).'-*',
+                null, $this->columnStart => (string) ($this->columnStart + 1),
+                default => ($this->columnStart + 1).'-'.($this->columnEnd + 1),
+            };
+        }
+
+        $selection = ($this->rowStart + 1).','.($this->columnStart + 1);
+
         return match (true) {
-            -1 === $this->rowStart => null,
-            null === $this->rowEnd => -1,
-            default => $this->rowEnd - $this->rowStart + 1,
+            $this->columnEnd === -1 => $selection.'-*',
+            $this->rowStart === $this->rowEnd && $this->columnStart === $this->columnEnd => $selection,
+            default => $selection.'-'.(($this->rowEnd ?? 0) + 1).','.(($this->columnEnd ?? 0) + 1),
         };
     }
 
@@ -58,6 +73,15 @@ final class Selection
             null === $this->columnEnd => null,
             default => range($this->columnStart, $this->columnEnd),
         };
+    }
+
+    public static function tryFromRow(string $selection): ?self
+    {
+        try {
+            return self::fromRow($selection);
+        } catch (FragmentNotFound) {
+            return null;
+        }
     }
 
     public static function fromRow(string $selection): self
@@ -72,6 +96,15 @@ final class Selection
         };
     }
 
+    public static function tryFromColumn(string $selection): ?self
+    {
+        try {
+            return self::fromColumn($selection);
+        } catch (FragmentNotFound) {
+            return null;
+        }
+    }
+
     public static function fromColumn(string $selection): self
     {
         [$start, $end] = self::parseRowColumnSelection($selection);
@@ -84,38 +117,13 @@ final class Selection
         };
     }
 
-    /**
-     * @return array{int<-1, max>, int|null|'*'}
-     */
-    private static function parseRowColumnSelection(string $selection): array
+    public static function tryFromCell(string $selection): ?self
     {
-        if (1 !== preg_match(self::REGEXP_ROWS_COLUMNS_SELECTION, $selection, $found)) {
-            return [-1, 0];
+        try {
+            return self::fromCell($selection);
+        } catch (FragmentNotFound) {
+            return null;
         }
-
-        $start = $found['start'];
-        $end = $found['end'] ?? null;
-        $start = filter_var($start, FILTER_VALIDATE_INT, ['options' => ['min_range' => 1]]);
-        if (false === $start) {
-            return [-1, 0];
-        }
-        --$start;
-
-        if (null === $end || '*' === $end) {
-            return [$start, $end];
-        }
-
-        $end = filter_var($end, FILTER_VALIDATE_INT, ['options' => ['min_range' => 1]]);
-        if (false === $end) {
-            return [-1, 0];
-        }
-        --$end;
-
-        if ($end <= $start) {
-            return [-1, 0];
-        }
-
-        return [$start, $end];
     }
 
     public static function fromCell(string $selection): self
@@ -156,5 +164,39 @@ final class Selection
         }
 
         return new self($cellStartRow, $cellEndRow, $cellStartCol, $cellEndCol,);
+    }
+
+    /**
+     * @return array{int<-1, max>, int|null|'*'}
+     */
+    private static function parseRowColumnSelection(string $selection): array
+    {
+        if (1 !== preg_match(self::REGEXP_ROWS_COLUMNS_SELECTION, $selection, $found)) {
+            return [-1, 0];
+        }
+
+        $start = $found['start'];
+        $end = $found['end'] ?? null;
+        $start = filter_var($start, FILTER_VALIDATE_INT, ['options' => ['min_range' => 1]]);
+        if (false === $start) {
+            return [-1, 0];
+        }
+        --$start;
+
+        if (null === $end || '*' === $end) {
+            return [$start, $end];
+        }
+
+        $end = filter_var($end, FILTER_VALIDATE_INT, ['options' => ['min_range' => 1]]);
+        if (false === $end) {
+            return [-1, 0];
+        }
+        --$end;
+
+        if ($end <= $start) {
+            return [-1, 0];
+        }
+
+        return [$start, $end];
     }
 }